requirements.txt
attrs==23.1.0
bcrypt==4.0.1
bidict==0.22.1
boolean-parser==0.1.4
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.6
commonmark==0.9.1
ConfigArgParse==0.15.2
cryptography==41.0.3
deepdiff==6.4.1
Deprecated==1.2.14
et-xmlfile==1.1.0
future==0.18.3
gitdb==4.0.10
GitPython==3.1.35
greenlet==2.0.2
idna==3.4
Jinja2==3.1.2
junos-eznc==2.6.7
lxml==4.9.3
MarkupSafe==2.1.3
mypy-extensions==0.4.4
napalm==4.1.0
ncclient==0.6.13
netaddr==0.8.0
netconan==0.12.3
netmiko==4.1.2
netutils==1.6.0
nornir==3.3.0
nornir-inspect==1.0.3
nornir-napalm==0.4.0
nornir-netmiko==1.0.0
nornir-scrapli==2023.7.30
nornir-utils==0.2.0
ntc-templates==3.5.0
numpy==1.25.2
openpyxl==3.1.2
ordered-set==4.1.0
pandas==1.1.5
paramiko==3.3.1
passlib==1.7.4
pybatfish==2023.5.12.784
pycparser==2.21
pyeapi==1.0.2
Pygments==2.16.1
PyNaCl==1.5.0
pynautobot==1.5.0
pyparsing==3.1.1
pyserial==3.5
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
pytricia==1.0.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-toolbelt==1.0.0
rich==12.6.0
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
scp==0.14.5
scrapli==2023.7.30
scrapli-cfg==2023.7.30
scrapli-community==2023.7.30
scrapli-netconf==2023.7.30
simplejson==3.19.1
six==1.16.0
smmap==5.0.0
SQLAlchemy==2.0.20
tabulate==0.9.0
tenacity==8.2.3
text-unidecode==1.3
textfsm==1.1.2
transitions==0.9.0
ttp==0.9.5
ttp-templates==0.3.5
typing_extensions==4.7.1
urllib3==1.26.16
-e git+https://github.com/veritas-sot/veritas.git@88ce7d5d1dbe6e33f92362b5cbdcfc046582a282#egg=veritas
wrapt==1.15.0
XlsxWriter==3.1.3
yamlordereddictloader==0.4.0
